3주차 상품 조회

+++ week02/src/main/java/com/example/week02/Order/OrderDto.java
@@ -14,7 +14,7 @@ import java.time.*;
 public record OrderDto(
 
         long userId,
-        int orderId,
+        long orderId,
         LocalDateTime orderDate,
         String delivery,
         String deliveryStatus)
